refactor (*): rename source dir
reorganize directory structure
