+++ hlib/graph/graph.py
