+++ hlib/graph/search.py
