+++ hlib/search/binary_search.py
