+++ hlib/tests/test_heap/test_min_binary_heap.py
@@ -1,45 +1,45 @@
 import unittest
 
-from data_structures.heap.max_binary_heap import MaxBinaryHeap
+from heap.min_binary_heap import MinBinaryHeap
 
 
 class TestStack(unittest.TestCase):
     def setUp(self): 
-        self.heap = MaxBinaryHeap()
+        self.heap = MinBinaryHeap()
 
     def test_push(self):
         self.heap.push(0)
         self.assertEqual(self.heap.peek(), 0)
         self.heap.push(-1)
-        self.assertEqual(self.heap.peek(), 0)
+        self.assertEqual(self.heap.peek(), -1)
         self.heap.push(-2)
-        self.assertEqual(self.heap.peek(), 0)
+        self.assertEqual(self.heap.peek(), -2)
         self.heap.push(1)
-        self.assertEqual(self.heap.peek(), 1)
+        self.assertEqual(self.heap.peek(), -2)
         self.heap.push(-3)
-        self.assertEqual(self.heap.peek(), 1)
+        self.assertEqual(self.heap.peek(), -3)
         self.heap.push(-4)
-        self.assertEqual(self.heap.peek(), 1)
+        self.assertEqual(self.heap.peek(), -4)
         self.heap.push(7)
-        self.assertEqual(self.heap.peek(), 7)
+        self.assertEqual(self.heap.peek(), -4)
 
     def test_push_increasing(self):
-        for i in range(0, 7):
-            self.heap.push(i)
-            self.assertEqual(self.heap.peek(), i)
-
-    def test_push_decreasing(self):
-        inputs = range(7, -1, -1)
+        inputs = range(0, 7)
         for i in inputs:
             self.heap.push(i)
             self.assertEqual(self.heap.peek(), inputs[0])
+
+    def test_push_decreasing(self):
+        for i in range(7, -1, -1):
+            self.heap.push(i)
+            self.assertEqual(self.heap.peek(), i)
 
     def test_pop(self):
         inputs = [0, -1, -2, 1, -3, -4, 7]
         for i in inputs:
             self.heap.push(i)
           
-        for i in reversed(sorted(inputs)):
+        for i in sorted(inputs):
             self.assertEqual(self.heap.pop(), i)
 
     def test_pop_increasing(self):
@@ -47,7 +47,7 @@ class TestStack(unittest.TestCase):
         for i in inputs:
             self.heap.push(i)
 
-        for i in reversed(inputs):
+        for i in inputs:
             self.assertEqual(self.heap.pop(), i)
 
     def test_pop_decreasing(self):
@@ -55,7 +55,7 @@ class TestStack(unittest.TestCase):
         for i in inputs:
             self.heap.push(i)
 
-        for i in inputs:
+        for i in reversed(inputs):
             self.assertEqual(self.heap.pop(), i)
 
     def test_pop_empty(self):
@@ -67,12 +67,12 @@ class TestStack(unittest.TestCase):
         self.assertRaises(Exception, self.heap.pop)
 
     def test_peek(self):
-        self.heap.push(-1)
-        self.assertEqual(self.heap.peek(), -1)
         self.heap.push(1)
         self.assertEqual(self.heap.peek(), 1)
+        self.heap.push(-1)
+        self.assertEqual(self.heap.peek(), -1)
         self.heap.push(0)
-        self.assertEqual(self.heap.peek(), 1)
+        self.assertEqual(self.heap.peek(), -1)
 
     def test_peek_empty(self):
         self.assertRaises(Exception, self.heap.peek)
